fix bugs (get token timing)
